+++ globus_cli/commands/__init__.py
@@ -1,0 +1,3 @@
+from globus_cli.commands.main import main
+
+__all__ = ['main']
