+++ globus_cli/commands/config_command/__init__.py
@@ -1,0 +1,5 @@
+from globus_cli.commands.config_command.commands import config_command
+
+__all__ = [
+    'config_command'
+]
