+++ globus_cli/commands/config_command/remove.py
@@ -2,19 +2,17 @@ import click
 
 from globus_cli.safeio import safeprint
 from globus_cli.parsing import common_options
+from globus_cli.config import get_config_obj
 
-from globus_cli.config_command.helpers import load_config
 
-
-@click.command('set', help='Set a value in the Globus Config')
+@click.command('remove', help='Remove a value from the Globus Config')
 @common_options(no_format_option=True)
 @click.argument('parameter', required=True)
-@click.argument('value', required=True)
-def set_command(value, parameter):
+def remove_command(parameter):
     """
-    Executor for `globus config set`
+    Executor for `globus config remove`
     """
-    conf = load_config()
+    conf = get_config_obj()
 
     section = 'general'
     if '.' in parameter:
@@ -23,8 +21,8 @@ def set_command(value, parameter):
     # ensure that the section exists
     if section not in conf:
         conf[section] = {}
-    # set the value for the given parameter
-    conf[section][parameter] = value
+    # remove the value for the given parameter
+    del conf[section][parameter]
 
     # write to disk
     safeprint('Writing updated config to {}'.format(conf.filename))
