+++ globus_cli/commands/config_command/set.py
@@ -2,18 +2,18 @@ import click
 
 from globus_cli.safeio import safeprint
 from globus_cli.parsing import common_options
+from globus_cli.config import get_config_obj
 
-from globus_cli.config_command.helpers import load_config
 
-
-@click.command('remove', help='Remove a value from the Globus Config')
+@click.command('set', help='Set a value in the Globus Config')
 @common_options(no_format_option=True)
 @click.argument('parameter', required=True)
-def remove_command(parameter):
+@click.argument('value', required=True)
+def set_command(value, parameter):
     """
-    Executor for `globus config remove`
+    Executor for `globus config set`
     """
-    conf = load_config()
+    conf = get_config_obj()
 
     section = 'general'
     if '.' in parameter:
@@ -22,8 +22,8 @@ def remove_command(parameter):
     # ensure that the section exists
     if section not in conf:
         conf[section] = {}
-    # remove the value for the given parameter
-    del conf[section][parameter]
+    # set the value for the given parameter
+    conf[section][parameter] = value
 
     # write to disk
     safeprint('Writing updated config to {}'.format(conf.filename))
