+++ globus_cli/commands/config_command/show.py
