--- globus_cli/config_command/__init__.py
@@ -1,5 +1,0 @@
-from globus_cli.config_command.commands import config_command
-
-__all__ = [
-    'config_command'
-]
