+++ globus_cli/helpers/__init__.py
@@ -1,12 +1,9 @@
-from globus_cli.helpers.identities import is_valid_identity_name
 from globus_cli.helpers.printing import (
     print_json_response, colon_formatted_print, print_table)
 from globus_cli.helpers.options import outformat_is_json, outformat_is_text
 
 
 __all__ = [
-    'is_valid_identity_name',
-
     'print_json_response', 'colon_formatted_print', 'print_table',
 
     'outformat_is_json', 'outformat_is_text'
