--- globus_cli/services/auth/__init__.py
@@ -1,10 +1,0 @@
-from globus_cli.services.auth.commands import auth_command
-from globus_cli.services.auth.helpers import (
-    lookup_identity_name, maybe_lookup_identity_id)
-
-
-__all__ = [
-    'auth_command',
-
-    'lookup_identity_name', 'maybe_lookup_identity_id'
-]
